Summarize conformance report by default with expandable details
Previously, every thing was visible at the same time. This made it
hard to get an overview of which scenarioes existed and the status of
each of them.

With a bit of "modern" HTML, provide expandable sections for every
part of the report. The role and scenario list is expanded by default.

The formatting is also changed so that the one-line scenario summary
also includes the status icon (but not status text). Similar
formatting changes have between done for other items to make it
prettier.

Signed-off-by: Niels Thykier <nit@globeteam.com>