DT-758 [1/?] Perform SI + TD status validation
Started on this while waiting for DT-767 as it was stalling my
progress on state related content. This is the changes I had time to
do on DT-758 in that time and it made sense to get those changes
merged to minimize the delta for future code. However, this is not
a completion of DT-758 (accordingly, the `[1/?]`).

Signed-off-by: Niels Thykier <nit@globeteam.com>