Refactor conformance scenario creation
Move the `ConformanceScenario` creation into the `AbstractComponantFactory`
and reduce visbility of `createModuleScenarioListBuilders`. This will enable
us to use different scenarion creation methods for different standards.

In the `ConformanceOrchestrator`, reduce the typing of
`scenariosByModuleName` to `List<ConformanceScenario>` from
`ArrayList<...>`. The actual list is not used externally and
internally the orchestrator will always just iterate over it via
stream or for-each loops. Given there is no "random access" accessing,
then any `List`-type will do. This in turn will enable more
flexibility for the scenarion building code to choose which list is
used.

The moved code is also slightly simplified because the code creating
the scenarios will not also populate the `_scenariosById`.