Expand on an error message and related comment
Users should not be able to trigger it since the JWS currently only
supports two key types relevant for a X.509 certificate. That is, they
would not be able to use the key themselves and follow the JWS
standard as it is right now.

However, X.509 potentially supports more key types, so a user can
create an X.509 certificate in PEM format. The code is there to catch
that (currently) theorectical case.