SD-1963: Use the RFC8785 approved Java library for JSON canonicalization
We received a report in products that a supporting tool were not
producing the correct checksums for TDs. Turns out that the library we
were using (same as in Conformance Framework) does not produce the
same checksum as the Java library listed RFC8785. This commit changes
the library to use the one used in RFC8785 to avoid future cases of
mismatches.